requirements.txt
pandas
numpy
scikit-learn
nltk
tensorflow
streamlit
vaderSentiment
Pillow